# service-contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: Digitaly Telemedicine API
  version: 0.1.0
  description: >
    Shared HTTP contract for the Digitaly telemedicine platform. The API is served
    under the global prefix /api, therefore every path below includes it.
servers:
  - url: /
...
paths:
...
  /api/patients/{patientId}/overview:
    get:
      tags: [records]
      operationId: getPatientOverview
      summary: Get patient profile, upcoming appointment, history and pre-consult
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientOverviewDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/appointments/{id}/pre-consult:
    get:
      tags: [appointments]
      operationId: getPreConsult
      summary: Get pre-consult answers for an appointment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
      responses:
        '200':
          description: Pre-consult answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PreConsultAnswerDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [appointments]
      operationId: putPreConsult
      summary: Create or replace pre-consult answers for an appointment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PreConsultAnswerInputDto'
      responses:
        '200':
          description: Saved pre-consult answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PreConsultAnswerDto'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/records/history:
    get:
      tags: [records]
      operationId: listConsultationHistory
      summary: List consultation history for a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Consultation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsultationHistoryItemDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/records:
    post:
      tags: [records]
      operationId: createMedicalRecord
      summary: Create a medical record for a consultation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMedicalRecordRequestDto'
      responses:
        '201':
          description: Medical record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalRecordDto'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Record already exists for this consultation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/records/{recordId}:
    get:
      tags: [records]
      operationId: getMedicalRecord
      summary: Get one medical record
      security:
        - bearerAuth: []
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Medical record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalRecordDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AppointmentId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Request could not be processed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
  schemas:
    UserRole:
      type: string
      enum: [doctor, patient]
    AppointmentStatus:
      type: string
      enum: [pending_code, confirmed, in_progress, completed, cancelled]
...
    ErrorResponseDto:
      type: object
      required:
        - statusCode
        - error
        - message
        - details
        - timestamp
        - path
        - correlationId
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: [array, 'null']
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        correlationId:
          type: string
...
    UserDto:
      type: object
      required: [id, name, email, role, specialty, crm, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        specialty:
          type: [string, 'null']
        crm:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...
    CounterpartDto:
      type: object
      required: [id, name, role, specialty]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        specialty:
          type: [string, 'null']
...
    PreConsultAnswerInputDto:
      type: object
      required: [questionKey, answer]
      properties:
        questionKey:
          type: string
        answer:
          type: string
    PreConsultAnswerDto:
      type: object
      required: [id, appointmentId, questionKey, answer, createdAt]
      properties:
        id:
          type: string
        appointmentId:
          type: string
        questionKey:
          type: string
        answer:
          type: string
        createdAt:
          type: string
          format: date-time
...
    AppointmentDto:
      type: object
      required: [id, patientId, doctorId, scheduledAt, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ConsultationHistoryItemDto:
      type: object
      required: [consultationId, appointmentId, startedAt, endedAt, doctor, diagnosis]
      properties:
        consultationId:
          type: string
        appointmentId:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: [string, 'null']
          format: date-time
        doctor:
          $ref: '#/components/schemas/CounterpartDto'
        diagnosis:
          type: [string, 'null']
    CreateMedicalRecordRequestDto:
      type: object
      required: [consultationId, patientId, notes]
      properties:
        consultationId:
          type: string
        patientId:
          type: string
        notes:
          type: string
        diagnosis:
          type: string
        prescriptions:
          type: array
          items:
            type: string
    MedicalRecordDto:
      type: object
      required:
        [id, patientId, consultationId, createdBy, notes, diagnosis, prescriptions, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        consultationId:
          type: string
        createdBy:
          type: string
        notes:
          type: string
        diagnosis:
          type: [string, 'null']
        prescriptions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    MedicalRecordSummaryDto:
      type: object
      required: [id, consultationId, createdAt, diagnosis]
      properties:
        id:
          type: string
        consultationId:
          type: string
        createdAt:
          type: string
          format: date-time
        diagnosis:
          type: [string, 'null']
    PatientOverviewDto:
      type: object
      required: [patient, upcomingAppointment, history, preConsult]
      properties:
        patient:
          $ref: '#/components/schemas/UserDto'
        upcomingAppointment:
          oneOf:
            - $ref: '#/components/schemas/AppointmentDto'
            - type: 'null'
        history:
          type: array
          items:
            $ref: '#/components/schemas/MedicalRecordSummaryDto'
        preConsult:
          type: array
          items:
            $ref: '#/components/schemas/PreConsultAnswerDto'
...